// src/types.ts
export type typeDefaultStyleObject = {
    styleBold: string;
    styleUnderline: string;
    styleTextSize: string;
};

export type typeWordNode = {
    text: string;
    styleBold: string;
    styleUnderline: string;
    styleTextSize: string;
};

export type typeResultCheckObject = {
    isFirstWordBold: boolean;
    isSecondWordUnderlined: boolean;
    thirdWordTextSize: string;
};